requirements.txt
playwright>=1.42.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.10
requests>=2.31.0
openai>=1.30.0
python-dotenv>=1.0.0
readability-lxml>=0.8.1
